+++ CHANGELOG.md
@@ -37,6 +37,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Replaced Sphinx with Mkdocs for documentation generation.
+- *BREAKING CHANGE:* Renamed `KokoroDeviceNames` to `KokoroDeviceTypes`.
 
 ### Deprecated